KAN-107 refactor: products 관련 타입 컨벤션 수정
